requirements.txt
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.2.2
xgboost==1.7.5
mlflow==2.3.1
pyyaml==6.0
joblib==1.2.0